requirements.txt
django==4.2.2

#phone number field
django-phonenumber-field[phonenumbers]==7.1.0